requirements.txt
numpy
pandas
scikit-learn
matplotlib
gensim
umap-learn
transformers